+++ src/OpenKNX/Hardware.cpp
@@ -229,6 +229,13 @@ namespace OpenKNX
     #else
         #pragma GCC error "No valid KNX UART interface defined (KNX_UART_NUM, KNX_UART_RX_PIN, KNX_UART_TX_PIN)"
     #endif
+
+        knx.bau().getDataLinkLayer()->getTPUart().registerReceivedFrame(
+            [](TPUart::Frame& tpFrame) {
+                // Process received frame
+                if (openknx.console.bcuDebug())
+                    openknx.logger.logWithPrefixAndValues("BCU<Debug>", "Received frame: %s", tpFrame.printFrame().c_str());
+            });
     }
 #endif